pyproject.toml
[tool.poetry]
name = "distributed-observability-tools"
version = "0.1.1"
description = "Comprehensive observability tools for distributed systems - tracing, correlation, and metrics"
authors = ["Tushar Khanka <tusharkhanka@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "distributed_observability"}]
repository = "https://github.com/tusharkhanka/distributed-observability-tools"
homepage = "https://github.com/tusharkhanka/distributed-observability-tools"
documentation = "https://github.com/tusharkhanka/distributed-observability-tools#readme"
keywords = ["observability", "tracing", "opentelemetry", "distributed-systems", "correlation-id", "fastapi", "microservices"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Distributed Computing",
]

[tool.poetry.dependencies]
python = "^3.8"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.21.0"
opentelemetry-instrumentation-fastapi = {version = "^0.42b0", optional = true}
opentelemetry-instrumentation-httpx = {version = "^0.42b0", optional = true}
opentelemetry-instrumentation-sqlalchemy = {version = "^0.42b0", optional = true}
opentelemetry-instrumentation-redis = {version = "^0.42b0", optional = true}
opentelemetry-instrumentation-botocore = {version = "^0.42b0", optional = true}
opentelemetry-instrumentation-grpc = {version = "^0.42b0", optional = true}
fastapi = {version = "^0.104.0", optional = true}
httpx = {version = "^0.25.0", optional = true}
celery = {version = "^5.3.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
pytest-asyncio = "^0.21.0"

[tool.poetry.extras]
fastapi = ["fastapi", "opentelemetry-instrumentation-fastapi"]
httpx = ["httpx", "opentelemetry-instrumentation-httpx"]
celery = ["celery"]
database = ["opentelemetry-instrumentation-sqlalchemy", "opentelemetry-instrumentation-redis"]
aws = ["opentelemetry-instrumentation-botocore"]
grpc = ["opentelemetry-instrumentation-grpc"]
all = [
    "fastapi", "httpx", "celery",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-instrumentation-redis",
    "opentelemetry-instrumentation-botocore",
    "opentelemetry-instrumentation-grpc"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.poetry.urls]
"Bug Reports" = "https://github.com/tusharkhanka/distributed-observability-tools/issues"
"Source" = "https://github.com/tusharkhanka/distributed-observability-tools"
"Documentation" = "https://github.com/tusharkhanka/distributed-observability-tools#readme"
"Changelog" = "https://github.com/tusharkhanka/distributed-observability-tools/releases"
